api/requirements.txt
Flask
Flask_SQLAlchemy
Flask_CORS
python-dotenv
PyMySQL
